Lift default transport detection to wallet logic
